Minor changes
i/ include/depth_node/defines.h passed roslint checks
ii/ fixed minor return issue in depth_filters

	modified:   pandora_vision_hole_detector/CMakeLists.txt
	modified:   pandora_vision_hole_detector/include/depth_node/defines.h
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/depth_filters.cpp
	modified:   pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp

+++ pandora_vision_hole_detector/src/hole_fusion_node/hole_fusion.cpp
@@ -300,8 +300,8 @@ namespace pandora_vision
 
     struct dirent* result = NULL;
 
-    long int nameMax = pathconf(wallPicturesPath.c_str(), _PC_NAME_MAX);
-    long int len = offsetof(struct dirent, d_name) + nameMax + 1;
+    int nameMax = pathconf(wallPicturesPath.c_str(), _PC_NAME_MAX);
+    int len = offsetof(struct dirent, d_name) + nameMax + 1;
     struct dirent *theDir = static_cast<struct dirent*>(malloc(len));
 
     DIR *directory;
